updated at 42

+++ srcs/ft_strlen.c
@@ -10,9 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(char *str)
+#include <stddef.h>
+
+size_t	ft_strlen(char *str)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (*str++)
